requirements.txt
fastapi
uvicorn
pygithub
google-genai
requests-oauthlib
python-dotenv
pyjwt[crypto]
redis